requirements.txt
streamlit
opencv-python
numpy
matplotlib
pillow